Refactor GPT-5.5 prompt variants by experiment flags
